requirements.txt
openai>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
reportlab>=4.0.0
click>=8.1.0
